[IMPROVEMENT] Compare tests against last commit results (#69)
This improvement changes the code so that we can compare against the latest master results instead of against a known version. This should make it easier to spot issues that are introduced by certain commits or PR's when the current repo already has some changes compared to the last version. Closes #31.