Re-add threading when starting new tests (#130)
When the platform starts a new test after finishing a previous one, it sleeps for 60 seconds.

When we had both platforms running in parallel this was ran in a separate process, but this process was removed when we reverted the parallelism. Because of this, the main thread now needs at least 60 seconds to complete, but GUnicorn only allows 30 seconds before killing the process. The end result is that the code is most likely not fully executed, and that the logs contain status code 504.