Use separate expiry time for HMAC
Instead of using the time that the HMAC was created as the expiry time,
use a separate expiry time in the future. This prevents a race condition
where these tests will sometimes fail if they run more than 1 second as
the HMAC would have expired by then.

The expiry time is set to one year from the time that the set up code
runs. That should hopefully be enough time to run the tests, seeing as,
if these tests run for more than a year, we probably have bigger
problems.