Make test_if_link_expired use correct expired HMAC
Currently test_if_link_expired uses a HMAC that is incorrect because it
uses a different expiry time than the one uses to generate the HMAC. On
top of this, it uses an expiry time in the future to test the response
when the HMAC is expired, when it should be using an expiry time in the
past.

This causes it to appear to still work because the complete_signup
request fails as expected, but that happens because the hash is
incorrect, not because the link is expired.

This commit fixes that bug to use the correct expiry time and HMAC
combination.